feat: added schema checking and validation to prevent duplicate file inclusion in database
