+++ src/schema.py
@@ -1,4 +1,5 @@
 import pandera as pa
+import pandas as pd
 from pandera.typing import Series
 
 class CompanyRevenue(pa.SchemaModel):
@@ -13,3 +14,8 @@ class CompanyRevenue(pa.SchemaModel):
         strict = True # make sure all specified columns are in the validated dataframe 
         # from_format = # data format before validation.
         # to_format = # data format to serialize into after validation.
+
+
+@pa.check_input(CompanyRevenue, lazy = False)
+def validate_schema_files(df: pd.DataFrame) -> pd.DataFrame:
+    return df
